+++ src/bootware.sh
@@ -1,4 +1,4 @@
-#!/usr/bin/env bash
+#!/usr/bin/env sh
 #
 # Bootstrap software installations with Ansible.
 
@@ -6,9 +6,8 @@
 #
 # Flags:
 #   -e: Exit immediately when a command pipeline fails.
-#   -o: Persist nonzero exit codes through a Bash pipe.
 #   -u: Throw an error when an unset variable is encountered.
-set -eou pipefail
+set -eu
 
 #######################################
 # Show CLI help information.
@@ -46,11 +45,10 @@ Options:
       --temp-key <FILE>           Path to SSH private key for one time connection
   -u, --url <URL>                 URL of playbook repository
       --user <USER>               Remote user login name
-      --windows                   Connect to a Windows host with SSH
 
 Ansible Options:
 EOF
-      if [[ -x "$(command -v ansible)" ]]; then
+      if [ -x "$(command -v ansible)" ]; then
         ansible --help
       fi
       ;;
@@ -143,7 +141,8 @@ Options:
 EOF
       ;;
     *)
-      error "No such usage option '${1}'"
+      log --stderr "No such usage option '${1}'."
+      exit 1
       ;;
   esac
 }
@@ -167,7 +166,7 @@ bootstrap() {
   local cmd='pull'
   local config_path="${BOOTWARE_CONFIG:-'/dev/null'}"
   local connection='local'
-  local extra_args=()
+  local extras=0
   local install_group
   local install_user
   local inventory='127.0.0.1,'
@@ -180,26 +179,22 @@ bootstrap() {
   local start_role
   local status
   local tags="${BOOTWARE_TAGS:-}"
-  local temp_ssh_args=(
-    "-o IdentitiesOnly=yes"
-    "-o LogLevel=ERROR"
-    "-o PreferredAuthentications=publickey,password"
-    "-o StrictHostKeyChecking=no"
-    "-o UserKnownHostsFile=/dev/null"
-  )
+  local temp_ssh_args='-o IdentitiesOnly=yes -o LogLevel=ERROR \
+-o PreferredAuthentications=publickey,password \
+-o StrictHostKeyChecking=no \
+-o UserKnownHostsFile=/dev/null'
   local url="${BOOTWARE_URL:-https://github.com/scruffaluff/bootware.git}"
-  local windows
 
   # Check if Ansible should ask for user password.
   #
   # Flags:
   #   -z: Check if the string is empty.
-  if [[ -z "${BOOTWARE_NOPASSWD:-}" ]]; then
+  if [ -z "${BOOTWARE_NOPASSWD:-}" ]; then
     ask_passwd='true'
   fi
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ] && [ "${extras}" -lt "${#}" ]; do
     case "${1}" in
       --ansible-config)
         export ANSIBLE_CONFIG="${2}"
@@ -278,28 +273,18 @@ bootstrap() {
         shift 2
         ;;
       --temp-key)
-        extra_args+=(
-          "--private-key"
-          "${2}"
-          "--ssh-extra-args"
-          "${temp_ssh_args[*]}"
-        )
+        set -- "$@" --private-key "${2}" --ssh-extra-args "${temp_ssh_args}"
         shift 2
+        extras="$((extras + 4))"
         ;;
       -u | --url)
         url="${2}"
         shift 2
         ;;
-      --windows)
-        ask_passwd=''
-        cmd='playbook'
-        connection='ssh'
-        windows='true'
-        shift 1
-        ;;
       *)
-        extra_args+=("${1}")
+        set -- "$@" "${1}"
         shift 1
+        extras="$((extras + 1))"
         ;;
     esac
   done
@@ -308,7 +293,7 @@ bootstrap() {
   #
   # Flags:
   #   -z: Check if the string is empty.
-  if [[ -z "${no_setup:-}" ]]; then
+  if [ -z "${no_setup:-}" ]; then
     setup
   fi
 
@@ -316,10 +301,11 @@ bootstrap() {
   #
   # Flags:
   #   -z: Check if the string is empty.
-  if [[ "${cmd}" == 'playbook' && -z "${playbook:-}" ]]; then
-    # Do not use long form --dry-run flag. It is not supported on MacOS.
+  if [ "${cmd}" = 'playbook' ] && [ -z "${playbook:-}" ]; then
+    # Do not use long form flags for mktemp. They are not supported on some
+    # systems.
     tmp_dir="$(mktemp -u)"
-    git clone --depth 1 "${url}" "${tmp_dir}" &> /dev/null
+    git clone --depth 1 "${url}" "${tmp_dir}" > /dev/null 2>&1
     playbook="${tmp_dir}/playbook.yaml"
   fi
 
@@ -327,61 +313,65 @@ bootstrap() {
   #
   # Flags:
   #   -n: Check if string is nonempty.
-  if [[ -n "${start_role:-}" ]]; then
+  if [ -n "${start_role:-}" ]; then
     repo_dir="$(dirname "${playbook}")"
     start_task="$(
       yq --exit-status '.[0].name' "${repo_dir}/ansible_collections/scruffaluff/bootware/roles/${start_role}/tasks/main.yaml"
     )"
-    extra_args+=('--start-at-task' "${start_task}")
+    set -- "$@" '--start-at-task' "${start_task}"
   fi
 
   # Convenience logic for using a single host without a trailing comma.
-  if [[ ! "${inventory}" =~ .*','.* ]]; then
-    inventory="${inventory},"
-  fi
+  case "${inventory}" in
+    *,*) ;;
+    *)
+      inventory="${inventory},"
+      ;;
+  esac
 
-  if [[ "${cmd}" == 'playbook' ]]; then
+  if [ "${cmd}" = 'playbook' ]; then
     ansible_config_path="$(dirname "${playbook}")/ansible.cfg"
-    if [[ -z "${ANSIBLE_CONFIG:-}" && -f "${ansible_config_path}" ]]; then
+    if [ -z "${ANSIBLE_CONFIG:-}" ] && [ -f "${ansible_config_path}" ]; then
       export ANSIBLE_CONFIG="${ansible_config_path}"
     fi
-    extra_args+=('--connection' "${connection}")
-  elif [[ "${cmd}" == 'pull' ]]; then
+    set -- "$@" '--connection' "${connection}"
+  elif [ "${cmd}" = 'pull' ]; then
     playbook="${BOOTWARE_PLAYBOOK:-playbook.yaml}"
-    extra_args+=('--url' "${url}")
+    set -- "$@" '--url' "${url}"
   fi
 
   find_config_path "${config_path}"
   config_path="${RET_VAL}"
-  if [[ "${EUID}" -ne 0 &&
-    -z "${become_method:-}" &&
-    "${inventory}" == '127.0.0.1,' ]]; then
-    become_method="$(find_super)"
-  fi
 
   log "Executing Ansible ${cmd}"
-  log 'Enter your user account password if prompted'
+  if [ -n "${ask_passwd:-}" ]; then
+    log 'Enter your user account password when prompted'
+  fi
 
+  # Do not quote extra_args. Otherwise extra_args will be interpreted as a
+  # single argument.
+  # shellcheck disable=SC2086
   until "ansible-${cmd}" \
     ${ask_passwd:+--ask-become-pass} \
     ${checkout:+--checkout "${checkout}"} \
-    ${install_group:+--extra-vars "group_id=${install_group}"} \
-    ${install_user:+--extra-vars "user_id=${install_user}"} \
+    --extra-vars "@${config_path}" \
     ${become_method:+--extra-vars "ansible_become_method=${become_method}"} \
     ${passwd:+--extra-vars "ansible_password=${passwd}"} \
-    ${port:+--extra-vars "ansible_ssh_port=${port}"} \
-    ${windows:+--extra-vars 'ansible_pkg_mgr=scoop'} \
     --extra-vars 'ansible_python_interpreter=auto_silent' \
-    ${windows:+--extra-vars 'ansible_shell_type=powershell'} \
-    --extra-vars "@${config_path}" \
+    ${port:+--extra-vars "ansible_ssh_port=${port}"} \
+    ${install_group:+--extra-vars "group_id=${install_group}"} \
+    ${install_user:+--extra-vars "user_id=${install_user}"} \
     --inventory "${inventory}" \
     ${tags:+--tags "${tags}"} \
     ${skip:+--skip-tags "${skip}"} \
-    ${extra_args:+"${extra_args[@]}"} \
+    "$@" \
     "${playbook}"; do
 
     status=$?
-    ((retries--)) && ((retries == 0)) && exit "${status}"
+    retries="$((retries - 1))"
+    if [ "${retries}" -eq 0 ]; then
+      exit "${status}"
+    fi
     printf "\nBootstrapping attempt failed with exit code %s." "${status}"
     printf "\nRetrying bootstrapping with %s attempts left.\n" "${retries}"
     sleep 4
@@ -390,10 +380,6 @@ bootstrap() {
 
 #######################################
 # Subcommand to generate or download Bootware configuration file.
-# Globals:
-#   HOME
-# Arguments:
-#   Parent directory of Bootware script.
 # Outputs:
 #   Writes status information to stdout.
 #######################################
@@ -401,7 +387,7 @@ config() {
   local src_url dst_file="${HOME}/.bootware/config.yaml" empty_cfg
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       -d | --dest)
         dst_file="${2}"
@@ -420,31 +406,27 @@ config() {
         shift 2
         ;;
       *)
-        error_usage "No such option '${1}'" 'config'
+        log --stderr "error: No such option '${1}'."
+        log --stderr "Run 'bootware config --help' for usage."
+        exit 2
         ;;
     esac
   done
 
-  # Do not use long form --parents flag for mkdir. It is not supported on MacOS.
+  # Do not use long form flags for mkdir. They are not supported on some
+  # systems.
   mkdir -p "$(dirname "${dst_file}")"
 
   # Check if empty configuration file should be generated.
   #
   # Flags:
   #   -z: Check if the string is empty.
-  if [[ "${empty_cfg:-}" == 'true' || -z "${src_url:-}" ]]; then
+  if [ "${empty_cfg:-}" = 'true' ] || [ -z "${src_url:-}" ]; then
     log "Writing empty configuration file to ${dst_file}"
     printf 'super_passwordless: false' > "${dst_file}"
   else
     log "Downloading configuration file to ${dst_file}"
-
-    # Download configuration file.
-    #
-    # Flags:
-    #   -L: Follow redirect request.
-    #   -S: Show errors.
-    #   -f: Use archive file. Must be third flag.
-    curl -LSfs "${src_url}" --output "${dst_file}"
+    fetch --dest "${dst_file}" "${src_url}"
   fi
 }
 
@@ -459,53 +441,79 @@ config() {
 #   Super user elevation command.
 #######################################
 dnf_check_update() {
-  local code
-  ${1:+"${1}"} dnf check-update || {
+  local code super="${1:-}"
+  ${super:+"${super}"} dnf check-update || {
     code="$?"
-    [[ "${code}" -eq 100 ]] && return 0
+    [ "${code}" -eq 100 ] && return 0
     return "${code}"
   }
 }
 
 #######################################
-# Print error message and exit script with error code.
-# Outputs:
-#   Writes error message to stderr.
+# Perform network request.
 #######################################
-error() {
-  local bold_red='\033[1;31m' default='\033[0m'
-  # Flags:
-  #   -t <FD>: Check if file descriptor is a terminal.
-  if [[ -t 2 ]]; then
-    printf "${bold_red}error${default}: %s\n" "${1}" >&2
-  else
-    printf "error: %s\n" "${1}" >&2
-  fi
-  exit 1
-}
+fetch() {
+  local dst_file='-' mode='' super='' url=''
 
-#######################################
-# Print error message and exit script with usage error code.
-# Outputs:
-#   Writes error message to stderr.
-#######################################
-error_usage() {
-  local bold_red='\033[1;31m' default='\033[0m'
+  # Parse command line arguments.
+  while [ "${#}" -gt 0 ]; do
+    case "${1}" in
+      -d | --dest)
+        dst_file="${2}"
+        shift 2
+        ;;
+      -m | --mode)
+        mode="${2}"
+        shift 2
+        ;;
+      -s | --super)
+        super="${2}"
+        shift 2
+        ;;
+      *)
+        url="${1}"
+        shift 1
+        ;;
+    esac
+  done
+
+  # Create parent directory if it does not exist.
+  #
   # Flags:
-  #   -t <FD>: Check if file descriptor is a terminal.
-  if [[ -t 2 ]]; then
-    printf "${bold_red}error${default}: %s\n" "${1}" >&2
-  else
-    printf "error: %s\n" "${1}" >&2
+  #   -p: Make parent directories if necessary.
+  if [ "${dst_file}" != '-' ]; then
+    ${super:+"${super}"} mkdir -p "$(dirname "${dst_file}")"
   fi
-  printf "Run 'bootware %s--help' for usage.\n" "${2:+${2} }" >&2
-  exit 2
+
+  # Download with Curl or Wget.
+  #
+  # Flags:
+  #   -O <PATH>: Save download to path.
+  #   -q: Hide log output.
+  #   -v: Only show file path of command.
+  #   -x: Check if file exists and execute permission is granted.
+  if [ -x "$(command -v curl)" ]; then
+    ${super:+"${super}"} curl --fail --location --show-error --silent --output \
+      "${dst_file}" "${url}"
+  elif [ -x "$(command -v wget)" ]; then
+    ${super:+"${super}"} wget -q -O "${dst_file}" "${url}"
+  else
+    log --stderr 'error: Unable to find a network file downloader.'
+    log --stderr 'Install curl, https://curl.se, manually before continuing.'
+    exit 1
+  fi
+
+  # Change file permissions if chmod parameter was passed.
+  #
+  # Flags:
+  #   -n: Check if string has nonzero length.
+  if [ -n "${mode:-}" ]; then
+    ${super:+"${super}"} chmod "${mode}" "${dst_file}"
+  fi
 }
 
 #######################################
 # Find path of Bootware configuration file.
-# Globals:
-#   HOME
 # Arguments:
 #   User supplied configuration path.
 # Outputs:
@@ -514,15 +522,17 @@ error_usage() {
 #   Configuration file path.
 #######################################
 find_config_path() {
+  local path="${1:-}"
+
   # Flags:
   #   -f: Check if file exists and is a regular file.
   #   -n: Check if string is nonempty.
   #   -v: Only show file path of command.
-  if [[ -f "${1:-}" ]]; then
-    RET_VAL="${1}"
-  elif [[ -n "${BOOTWARE_CONFIG:-}" ]]; then
+  if [ -f "${path}" ]; then
+    RET_VAL="${path}"
+  elif [ -n "${BOOTWARE_CONFIG:-}" ]; then
     RET_VAL="${BOOTWARE_CONFIG}"
-  elif [[ -f "${HOME}/.bootware/config.yaml" ]]; then
+  elif [ -f "${HOME}/.bootware/config.yaml" ]; then
     RET_VAL="${HOME}/.bootware/config.yaml"
   else
     log 'Unable to find Bootware configuration file.'
@@ -540,21 +550,28 @@ find_super() {
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ -x "$(command -v sudo)" ]]; then
-    echo 'sudo'
-  elif [[ -x "$(command -v doas)" ]]; then
+  if [ -x "$(command -v doas)" ]; then
     echo 'doas'
+  elif [ -x "$(command -v sudo)" ]; then
+    echo 'sudo'
   else
-    error 'Unable to find a command for super user elevation'
+    log --stderr 'Unable to find a command for super user elevation.'
+    exit 1
   fi
 }
 
 #######################################
 # Get full normalized path for file.
+#
 # Alternative to realpath command, since it is not built into MacOS.
+#
+# Arguments:
+#   File system path.
+# Outputs:
+#   Full file system path.
 #######################################
 fullpath() {
-  local working_dir
+  local path="${1}" working_dir
 
   # Flags:
   #   -P: Resolve any symbolic links in the path.
@@ -569,16 +586,12 @@ fullpath() {
 #   Super user elevation command.
 #######################################
 install_yq() {
-  local arch os_type url version
+  local arch os url super="${1:-}" version
 
-  # Do not use long form --kernel-name or --machine flags for uname. They are
-  # not supported on MacOS.
-  arch="$(uname -m)"
-  arch="${arch/#x86_64/amd64}"
-  arch="${arch/%x64/amd64}"
-  arch="${arch/#aarch64/arm64}"
-  arch="${arch/%arm/arm64}"
-  os_type="$(uname -s)"
+  # Do not use long form flags for uname. They are not supported on some
+  # systems.
+  arch="$(uname -m | sed 's/x86_64/amd64/;s/x64/amd64/;s/aarch64/arm64/')"
+  os="$(uname -s)"
 
   # Get latest YQ version.
   #
@@ -589,29 +602,49 @@ install_yq() {
   #   -o <FILE>: Save output to file.
   #   -s: Disable progress bars.
   version="$(
-    curl -LSfs https://formulae.brew.sh/api/formula/yq.json |
+    fetch https://formulae.brew.sh/api/formula/yq.json |
       jq --exit-status --raw-output .versions.stable
   )"
-  url="https://github.com/mikefarah/yq/releases/download/v${version}/yq_${os_type}_${arch}"
-
-  # Do not quote the outer super parameter expansion. Shell will error due to be
-  # being unable to find the "" command.
-  ${1:+"${1}"} curl -LSfs "${url}" --output /usr/local/bin/yq
-  ${1:+"${1}"} chmod 755 /usr/local/bin/yq
+  url="https://github.com/mikefarah/yq/releases/download/v${version}/yq_${os}_${arch}"
+  fetch --dest /usr/local/bin/yq --mode 755 --super "${super}" "${url}"
 }
 
 #######################################
-# Print log message to stdout if logging is enabled.
+# Print message if error or logging is enabled.
+# Arguments:
+#   Message to print.
 # Globals:
 #   BOOTWARE_NOLOG
 # Outputs:
-#   Log message to stdout.
+#   Message argument.
 #######################################
 log() {
+  local file='1' newline="\n" text=''
+
+  # Parse command line arguments.
+  while [ "${#}" -gt 0 ]; do
+    case "${1}" in
+      -e | --stderr)
+        file='2'
+        shift 1
+        ;;
+      -n | --no-newline)
+        newline=''
+        shift 1
+        ;;
+      *)
+        text="${text}${1}"
+        shift 1
+        ;;
+    esac
+  done
+
+  # Print if error or using quiet configuration.
+  #
   # Flags:
-  #   -z: Check if the string is empty.
-  if [[ -z "${BOOTWARE_NOLOG:-}" ]]; then
-    echo "$@"
+  #   -z: Check if string has zero length.
+  if [ -z "${BOOTWARE_NOLOG:-}" ] || [ "${file}" = '2' ]; then
+    printf "%s${newline}" "${text}" >&"${file}"
   fi
 }
 
@@ -624,7 +657,7 @@ roles() {
   local url="${BOOTWARE_URL:-https://github.com/scruffaluff/bootware.git}"
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       -h | --help)
         usage 'roles'
@@ -639,7 +672,9 @@ roles() {
         shift 2
         ;;
       *)
-        error_usage "No such option '${1}'" 'roles'
+        log --stderr "error: No such option '${1}'."
+        log --stderr "Run 'bootware roles --help' for usage."
+        exit 2
         ;;
     esac
   done
@@ -647,12 +682,12 @@ roles() {
   # Do not use long form --dry-run flag for mktemp. It is not supported on
   # MacOS.
   tmp_dir="$(mktemp -u)"
-  git clone --depth 1 "${url}" "${tmp_dir}" &> /dev/null
+  git clone --depth 1 "${url}" "${tmp_dir}" > /dev/null 2>&1
 
   # Flags:
   #   -n: Check if string is nonempty.
-  if [[ -n "${tags:-}" ]]; then
-    contains="(map(. == \"${tags//,/\") | any) or (map(. == \"}\") | any)"
+  if [ -n "${tags:-}" ]; then
+    contains="(map(. == \"$(echo "${tags}" | sed 's/,/\") | any) or (map(. == \"/g')\") | any)"
     filter=".[0].tasks[] | select(.tags | (${contains}))"
   else
     filter='.[0].tasks[]'
@@ -666,30 +701,32 @@ roles() {
 # Subcommand to configure bootstrapping services and utilities.
 #######################################
 setup() {
-  local collections collection_status os_type tmp_dir super=''
+  local collections collection_status os tmp_dir super=''
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       -h | --help)
         usage 'setup'
         exit 0
         ;;
       *)
-        error_usage "No such option '${1}'" 'setup'
+        log --stderr "error: No such option '${1}'."
+        log --stderr "Run 'bootware setup --help' for usage."
+        exit 2
         ;;
     esac
   done
 
   # Check if user is not root.
-  if [[ "${EUID}" -ne 0 ]]; then
+  if [ "$(id -u)" -ne 0 ]; then
     super="$(find_super)"
   fi
 
   # Do not use long form --kernel-name flag for uname. It is not supported on
   # MacOS.
-  os_type="$(uname -s)"
-  case "${os_type}" in
+  os="$(uname -s)"
+  case "${os}" in
     Darwin)
       setup_macos
       ;;
@@ -700,16 +737,20 @@ setup() {
       setup_linux "${super}"
       ;;
     *)
-      error "Operating system ${os_type} is not supported"
+      log --stderr "Operating system '${os}' is not supported."
+      exit 1
       ;;
   esac
 
-  collections=('chocolatey.chocolatey' 'community.general' 'community.windows')
-  for collection in "${collections[@]}"; do
+  collections='chocolatey.chocolatey community.general community.windows'
+  for collection in ${collections}; do
     collection_status="$(ansible-galaxy collection list "${collection}" 2>&1)"
-    if [[ "${collection_status}" =~ 'unable to find' ]]; then
-      ansible-galaxy collection install "${collection}"
-    fi
+    case "${collection_status}" in
+      *'unable to find'*)
+        ansible-galaxy collection install "${collection}"
+        ;;
+      *) ;;
+    esac
   done
 }
 
@@ -719,36 +760,38 @@ setup() {
 #   Super user elevation command.
 #######################################
 setup_alpine() {
+  local super="${1:-}"
+
   # Install dependencies for Bootware.
   #
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
-    ${1:+"${1}"} apk update
-    ${1:+"${1}"} apk add ansible
+    ${super:+"${super}"} apk update
+    ${super:+"${super}"} apk add ansible
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
-    ${1:+"${1}"} apk update
-    ${1:+"${1}"} apk add curl
+    ${super:+"${super}"} apk update
+    ${super:+"${super}"} apk add curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
-    ${1:+"${1}"} apk update
-    ${1:+"${1}"} apk add git
+    ${super:+"${super}"} apk update
+    ${super:+"${super}"} apk add git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
-    ${1:+"${1}"} apk update
-    ${1:+"${1}"} apk add jq
+    ${super:+"${super}"} apk update
+    ${super:+"${super}"} apk add jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     install_yq "${1}"
   fi
@@ -760,42 +803,42 @@ setup_alpine() {
 #   Super user elevation command.
 #######################################
 setup_arch() {
-  local tmp_dir
+  local super="${1:-}" tmp_dir
 
   # Install dependencies for Bootware.
   #
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
     # Installing Ansible via Python causes pacman conflicts with AWSCLI.
-    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
-    ${1:+"${1}"} pacman --noconfirm --sync ansible
+    ${super:+"${super}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${super:+"${super}"} pacman --noconfirm --sync ansible
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
-    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
-    ${1:+"${1}"} pacman --noconfirm --sync curl
+    ${super:+"${super}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${super:+"${super}"} pacman --noconfirm --sync curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
-    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
-    ${1:+"${1}"} pacman --noconfirm --sync git
+    ${super:+"${super}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${super:+"${super}"} pacman --noconfirm --sync git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
-    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
-    ${1:+"${1}"} pacman --noconfirm --sync jq
+    ${super:+"${super}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${super:+"${super}"} pacman --noconfirm --sync jq
   fi
 
-  if [[ ! -x "$(command -v yay)" ]]; then
+  if [ ! -x "$(command -v yay)" ]; then
     log 'Installing Yay package manager'
-    ${1:+"${1}"} pacman --noconfirm --refresh --sync --sysupgrade
-    ${1:+"${1}"} pacman --noconfirm --sync base-devel
+    ${super:+"${super}"} pacman --noconfirm --refresh --sync --sysupgrade
+    ${super:+"${super}"} pacman --noconfirm --sync base-devel
 
     tmp_dir="$(mktemp --dry-run)"
     git clone --depth 1 'https://aur.archlinux.org/yay.git' "${tmp_dir}"
@@ -803,7 +846,7 @@ setup_arch() {
     yay --noconfirm --refresh --sync --sysupgrade
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     install_yq "${1}"
   fi
@@ -815,6 +858,8 @@ setup_arch() {
 #   Super user elevation command.
 #######################################
 setup_debian() {
+  local super="${1:-}"
+
   # Avoid APT interactively requesting to configure tzdata.
   export DEBIAN_FRONTEND='noninteractive'
 
@@ -823,33 +868,33 @@ setup_debian() {
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     # Install Ansible with Python3 since most package managers provide an old
     # version of Ansible.
     log 'Installing Ansible'
-    ${1:+"${1}"} apt-get --quiet update
-    ${1:+"${1}"} apt-get --quiet install --yes ansible
+    ${super:+"${super}"} apt-get --quiet update
+    ${super:+"${super}"} apt-get --quiet install --yes ansible
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
-    ${1:+"${1}"} apt-get --quiet update
-    ${1:+"${1}"} apt-get --quiet install --yes curl
+    ${super:+"${super}"} apt-get --quiet update
+    ${super:+"${super}"} apt-get --quiet install --yes curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
-    ${1:+"${1}"} apt-get --quiet update
-    ${1:+"${1}"} apt-get --quiet install --yes git
+    ${super:+"${super}"} apt-get --quiet update
+    ${super:+"${super}"} apt-get --quiet install --yes git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
-    ${1:+"${1}"} apt-get --quiet update
-    ${1:+"${1}"} apt-get --quiet install --yes jq
+    ${super:+"${super}"} apt-get --quiet update
+    ${super:+"${super}"} apt-get --quiet install --yes jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     install_yq "${1}"
   fi
@@ -861,38 +906,40 @@ setup_debian() {
 #   Super user elevation command.
 #######################################
 setup_fedora() {
+  local super="${1:-}"
+
   # Install dependencies for Bootware.
   #
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
     # Installing Ansible via Python causes issues installing remote DNF packages
     # with Ansible.
     dnf_check_update "${1}"
-    ${1:+"${1}"} dnf install --assumeyes ansible
+    ${super:+"${super}"} dnf install --assumeyes ansible
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
     dnf_check_update "${1}"
-    ${1:+"${1}"} dnf install --assumeyes curl
+    ${super:+"${super}"} dnf install --assumeyes curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
     dnf_check_update "${1}"
-    ${1:+"${1}"} dnf install --assumeyes git
+    ${super:+"${super}"} dnf install --assumeyes git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
     dnf_check_update "${1}"
-    ${1:+"${1}"} dnf install --assumeyes jq
+    ${super:+"${super}"} dnf install --assumeyes jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     install_yq "${1}"
   fi
@@ -904,36 +951,37 @@ setup_fedora() {
 #   Super user elevation command.
 #######################################
 setup_freebsd() {
-  local ansible_package
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  local ansible_package super="${1:-}"
+
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
-    ${1:+"${1}"} pkg update
+    ${super:+"${super}"} pkg update
 
     ansible_package="$(
       pkg search --quiet --regex 'py[0-9]+-ansible-[^A-Za-z]'
     )"
-    ${1:+"${1}"} pkg install --yes "${ansible_package}"
+    ${super:+"${super}"} pkg install --yes "${ansible_package}"
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
-    ${1:+"${1}"} pkg update
-    ${1:+"${1}"} pkg install --yes curl
+    ${super:+"${super}"} pkg update
+    ${super:+"${super}"} pkg install --yes curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
-    ${1:+"${1}"} pkg update
-    ${1:+"${1}"} pkg install --yes git
+    ${super:+"${super}"} pkg update
+    ${super:+"${super}"} pkg install --yes git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
-    ${1:+"${1}"} pkg update
-    ${1:+"${1}"} pkg install --yes jq
+    ${super:+"${super}"} pkg update
+    ${super:+"${super}"} pkg install --yes jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     install_yq "${1}"
   fi
@@ -945,23 +993,26 @@ setup_freebsd() {
 #   Super user elevation command.
 #######################################
 setup_linux() {
+  local super="${1:-}"
+
   # Install dependencies for Bootware base on available package manager.
   #
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ -x "$(command -v apk)" ]]; then
-    setup_alpine "${1}"
-  elif [[ -x "$(command -v pacman)" ]]; then
-    setup_arch "${1}"
-  elif [[ -x "$(command -v apt-get)" ]]; then
-    setup_debian "${1}"
-  elif [[ -x "$(command -v dnf)" ]]; then
-    setup_fedora "${1}"
-  elif [[ -x "$(command -v zypper)" ]]; then
-    setup_suse "${1}"
+  if [ -x "$(command -v apk)" ]; then
+    setup_alpine "${super}"
+  elif [ -x "$(command -v pacman)" ]; then
+    setup_arch "${super}"
+  elif [ -x "$(command -v apt-get)" ]; then
+    setup_debian "${super}"
+  elif [ -x "$(command -v dnf)" ]; then
+    setup_fedora "${super}"
+  elif [ -x "$(command -v zypper)" ]; then
+    setup_suse "${super}"
   else
-    error 'Unable to find supported package manager'
+    log --stderr 'Unable to find supported package manager.'
+    exit 1
   fi
 }
 
@@ -977,7 +1028,7 @@ setup_macos() {
   # Homebrew depends on the XCode command line tools.
   # Flags:
   #   -p: Print path to active developer directory.
-  if ! xcode-select -p &> /dev/null; then
+  if ! xcode-select -p > /dev/null 2>&1; then
     log 'Installing command line tools for XCode'
     sudo xcode-select --install
   fi
@@ -990,28 +1041,29 @@ setup_macos() {
   #   -f: Fail silently on server errors.
   #   -s: Disable progress bars.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v brew)" ]]; then
+  if [ ! -x "$(command -v brew)" ]; then
     log 'Installing Homebrew'
-    curl -LSfs 'https://raw.githubusercontent.com/Homebrew/install/master/install.sh' | bash
+    fetch 'https://raw.githubusercontent.com/Homebrew/install/master/install.sh' |
+      bash
     brew analytics off
   fi
 
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
     brew install ansible
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
     brew install git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
     brew install jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
     brew install yq
   fi
@@ -1023,38 +1075,40 @@ setup_macos() {
 #   Super user elevation command.
 #######################################
 setup_suse() {
+  local super="${1:-}"
+
   # Install dependencies for Bootware.
   #
   # Flags:
   #   -v: Only show file path of command.
   #   -x: Check if file exists and execute permission is granted.
-  if [[ ! -x "$(command -v ansible)" ]]; then
+  if [ ! -x "$(command -v ansible)" ]; then
     log 'Installing Ansible'
-    ${1:+"${1}"} zypper update --no-confirm
-    ${1:+"${1}"} zypper install --no-confirm ansible
+    ${super:+"${super}"} zypper update --no-confirm
+    ${super:+"${super}"} zypper install --no-confirm ansible
   fi
 
-  if [[ ! -x "$(command -v curl)" ]]; then
+  if [ ! -x "$(command -v curl)" ]; then
     log 'Installing Curl'
-    ${1:+"${1}"} zypper update --no-confirm
-    ${1:+"${1}"} zypper install --no-confirm curl
+    ${super:+"${super}"} zypper update --no-confirm
+    ${super:+"${super}"} zypper install --no-confirm curl
   fi
 
-  if [[ ! -x "$(command -v git)" ]]; then
+  if [ ! -x "$(command -v git)" ]; then
     log 'Installing Git'
-    ${1:+"${1}"} zypper update --no-confirm
-    ${1:+"${1}"} zypper install --no-confirm git
+    ${super:+"${super}"} zypper update --no-confirm
+    ${super:+"${super}"} zypper install --no-confirm git
   fi
 
-  if [[ ! -x "$(command -v jq)" ]]; then
+  if [ ! -x "$(command -v jq)" ]; then
     log 'Installing JQ'
-    ${1:+"${1}"} zypper update --no-confirm
-    ${1:+"${1}"} zypper install --no-confirm jq
+    ${super:+"${super}"} zypper update --no-confirm
+    ${super:+"${super}"} zypper install --no-confirm jq
   fi
 
-  if [[ ! -x "$(command -v yq)" ]]; then
+  if [ ! -x "$(command -v yq)" ]; then
     log 'Installing YQ'
-    install_yq "${1}"
+    install_yq "${super}"
   fi
 }
 
@@ -1067,14 +1121,16 @@ uninstall() {
   local dst_file super=''
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       -h | --help)
         usage 'uninstall'
         exit 0
         ;;
       *)
-        error_usage "No such option '${1}'" 'uninstall'
+        log --stderr "error: No such option '${1}'."
+        log --stderr "Run 'bootware uninstall --help' for usage."
+        exit 2
         ;;
     esac
   done
@@ -1085,7 +1141,7 @@ uninstall() {
   #
   # Flags:
   #   -w: Check if file exists and it writable.
-  if [[ ! -w "${dst_file}" && "${EUID}" -ne 0 ]]; then
+  if [ ! -w "${dst_file}" ] && [ "$(id -u)" -ne 0 ]; then
     super="$(find_super)"
   fi
 
@@ -1107,7 +1163,7 @@ update() {
   local dst_file src_url super='' user_install version='main'
 
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       -h | --help)
         usage 'update'
@@ -1118,7 +1174,9 @@ update() {
         shift 2
         ;;
       *)
-        error_usage "No such option '${1}'" 'update'
+        log --stderr "error: No such option '${1}'."
+        log --stderr "Run 'bootware update --help' for usage."
+        exit 2
         ;;
     esac
   done
@@ -1130,19 +1188,14 @@ update() {
   #
   # Flags:
   #   -w: Check if file exists and it writable.
-  if [[ ! -w "${dst_file}" && "${EUID}" -ne 0 ]]; then
+  if [ ! -w "${dst_file}" ] && [ "$(id -u)" -ne 0 ]; then
     super="$(find_super)"
-  elif [[ -w "${dst_file}" && "${EUID}" -ne 0 ]]; then
+  elif [ -w "${dst_file}" ] && [ "$(id -u)" -ne 0 ]; then
     user_install='true'
   fi
 
   log 'Updating Bootware'
-
-  # Do not quote the outer super parameter expansion. Shell will error due to be
-  # being unable to find the "" command.
-  ${super:+"${super}"} curl -LSfs "${src_url}" --output "${dst_file}"
-  ${super:+"${super}"} chmod 755 "${dst_file}"
-
+  fetch --dest "${dst_file}" --mode 755 --super "${super}" "${src_url}"
   update_completions "${super}" "${user_install:-}" "${version}"
   log "Updated to version $(bootware --version)"
 }
@@ -1155,56 +1208,44 @@ update() {
 #   GitHub version reference.
 #######################################
 update_completions() {
-  local brew_prefix os_type
-  local repo_url="https://raw.githubusercontent.com/scruffaluff/bootware/${3}"
+  local brew_prefix global_="${2}" os super="${1}" version="${3}"
+  local repo_url="https://raw.githubusercontent.com/scruffaluff/bootware/${version}"
   local bash_url="${repo_url}/src/completion/bootware.bash"
   local fish_url="${repo_url}/src/completion/bootware.fish"
 
   # Flags:
   #  -z: Check if the string is empty.
-  if [[ -z "${2:-}" ]]; then
-    if [[ "$(uname -m)" == 'arm64' ]]; then
+  if [ -z "${global_}" ]; then
+    if [ "$(uname -m)" = 'arm64' ]; then
       brew_prefix='/opt/homebrew'
     else
       brew_prefix='/usr/local'
     fi
-    os_type="$(uname -s)"
+    os="$(uname -s)"
 
     # Do not use long form --parents flag for mkdir. It is not supported on
     # MacOS.
-    if [[ "${os_type}" == 'Darwin' ]]; then
-      ${1:+"${1}"} mkdir -p "${brew_prefix}/share/bash-completion/completions"
-      ${1:+"${1}"} curl -LSfs "${bash_url}" --output "${brew_prefix}/share/bash-completion/completions/bootware"
-      ${1:+"${1}"} chmod 644 "${brew_prefix}/share/bash-completion/completions/bootware"
-
-      ${1:+"${1}"} mkdir -p "${brew_prefix}/etc/fish/completions"
-      ${1:+"${1}"} curl -LSfs "${fish_url}" --output "${brew_prefix}/etc/fish/completions/bootware.fish"
-      ${1:+"${1}"} chmod 644 "${brew_prefix}/etc/fish/completions/bootware.fish"
-    elif [[ "${os_type}" == 'FreeBSD' ]]; then
-      ${1:+"${1}"} mkdir -p '/usr/local/share/bash-completion/completions'
-      ${1:+"${1}"} curl -LSfs "${bash_url}" --output '/usr/local/share/bash-completion/completions/bootware'
-      ${1:+"${1}"} chmod 644 '/usr/local/share/bash-completion/completions/bootware'
-
-      ${1:+"${1}"} mkdir -p '/usr/local/etc/fish/completions'
-      ${1:+"${1}"} curl -LSfs "${fish_url}" --output '/usr/local/etc/fish/completions/bootware.fish'
-      ${1:+"${1}"} chmod 644 '/usr/local/etc/fish/completions/bootware.fish'
+    if [ "${os}" = 'Darwin' ]; then
+      fetch --dest "${brew_prefix}/share/bash-completion/completions/bootware" \
+        --mode 644 --super "${super}" "${bash_url}"
+      fetch --dest "${brew_prefix}/etc/fish/completions/bootware.fish" \
+        --mode 644 --super "${super}" "${fish_url}"
+    elif [ "${os}" = 'FreeBSD' ]; then
+      fetch --dest '/usr/local/share/bash-completion/completions/bootware' \
+        --mode 644 --super "${super}" "${bash_url}"
+      fetch --dest '/usr/local/etc/fish/completions/bootware.fish' \
+        --mode 644 --super "${super}" "${fish_url}"
     else
-      ${1:+"${1}"} mkdir -p '/usr/share/bash-completion/completions'
-      ${1:+"${1}"} curl -LSfs "${bash_url}" --output '/usr/share/bash-completion/completions/bootware'
-      ${1:+"${1}"} chmod 644 '/usr/share/bash-completion/completions/bootware'
-
-      ${1:+"${1}"} mkdir -p '/etc/fish/completions'
-      ${1:+"${1}"} curl -LSfs "${fish_url}" --output '/etc/fish/completions/bootware.fish'
-      ${1:+"${1}"} chmod 644 '/etc/fish/completions/bootware.fish'
+      fetch --dest '/usr/share/bash-completion/completions/bootware' \
+        --mode 644 --super "${super}" "${bash_url}"
+      fetch --dest '/etc/fish/completions/bootware.fish' \
+        --mode 644 --super "${super}" "${fish_url}"
     fi
   else
-    mkdir -p "${HOME}/.local/share/bash-completion/completions"
-    curl -LSfs "${bash_url}" --output "${HOME}/.local/share/bash-completion/completions/bootware"
-    chmod 644 "${HOME}/.local/share/bash-completion/completions/bootware"
-
-    mkdir -p "${HOME}/.config/fish/completions"
-    curl -LSfs "${fish_url}" --output "${HOME}/.config/fish/completions/bootware.fish"
-    chmod 644 "${HOME}/.config/fish/completions/bootware.fish"
+    fetch --dest "${HOME}/.local/share/bash-completion/completions/bootware" \
+      --mode 644 "${bash_url}"
+    fetch --dest "${HOME}/.config/fish/completions/bootware.fish" \
+      --mode 644 "${fish_url}"
   fi
 }
 
@@ -1222,7 +1263,7 @@ version() {
 #######################################
 main() {
   # Parse command line arguments.
-  while [[ "${#}" -gt 0 ]]; do
+  while [ "${#}" -gt 0 ]; do
     case "${1}" in
       --debug)
         set -o xtrace
@@ -1267,7 +1308,9 @@ main() {
         exit 0
         ;;
       *)
-        error_usage "No such subcommand or option '${1}'"
+        log --stderr "error: No such subcommand or option '${1}'."
+        log --stderr "Run 'bootware --help' for usage."
+        exit 2
         ;;
     esac
   done
@@ -1276,6 +1319,6 @@ main() {
 }
 
 # Add ability to selectively skip main function during test suite.
-if [[ -z "${BATS_SOURCE_ONLY:-}" ]]; then
+if [ -z "${BATS_SOURCE_ONLY:-}" ]; then
   main "$@"
 fi