+++ test/unit/bootstrap.test.ps1
@@ -23,7 +23,7 @@ Describe 'Bootstrap' {
     It 'Subcommand finds first task associated with role' {
         $Env:BOOTWARE_NOLOG = 1
         $Playbook = "$(Get-Location)\playbook.yaml"
-        $Expected = 'wsl bootware bootstrap --windows --config ' `
+        $Expected = 'wsl bootware bootstrap --no-passwd --config ' `
             + '/mnt/c/Users/Administrator/.bootware/config.yaml --inventory ' `
             + "192.48.16.0 --playbook $(WSLPath $Playbook) " `
             + "--private-key `$HOME/.ssh/bootware --skip none " `
@@ -37,7 +37,7 @@ Describe 'Bootstrap' {
 
     It 'Subcommand passes default arguments to WSL copy of Bootware' {
         $Env:BOOTWARE_NOLOG = 1
-        $Expected = 'wsl bootware bootstrap --windows --config ' `
+        $Expected = 'wsl bootware bootstrap --no-passwd --config ' `
             + '/mnt/c/Users/Administrator/.bootware/config.yaml --inventory ' `
             + '192.48.16.0 --playbook /mnt/c/Fake/path/repo/playbook.yaml ' `
             + "--private-key `$HOME/.ssh/bootware --skip none " `
@@ -50,7 +50,7 @@ Describe 'Bootstrap' {
 
     It 'Subcommand passes debug argument to WSL copy of Bootware' {
         $Env:BOOTWARE_NOLOG = 1
-        $Expected = 'wsl bootware bootstrap --windows --config ' `
+        $Expected = 'wsl bootware bootstrap --no-passwd --config ' `
             + '/mnt/c/Users/Administrator/.bootware/config.yaml --inventory ' `
             + '192.48.16.0 --playbook /mnt/c/Fake/path/repo/playbook.yaml ' `
             + "--private-key `$HOME/.ssh/bootware --skip python " `
@@ -64,7 +64,7 @@ Describe 'Bootstrap' {
 
     It 'Subcommand passes list arguments to WSL copy of Bootware' {
         $Env:BOOTWARE_NOLOG = 1
-        $Expected = 'wsl bootware bootstrap --windows --config ' `
+        $Expected = 'wsl bootware bootstrap --no-passwd --config ' `
             + '/mnt/c/Users/Administrator/.bootware/config.yaml --inventory ' `
             + '192.48.16.0 --playbook /mnt/c/Fake/path/repo/playbook.yaml ' `
             + "--private-key `$HOME/.ssh/bootware --skip rust " `
@@ -78,7 +78,7 @@ Describe 'Bootstrap' {
 
     It 'Subcommand passes extra arguments to WSL copy of Bootware' {
         $Env:BOOTWARE_NOLOG = 1
-        $Expected = 'wsl bootware bootstrap --windows --config ' `
+        $Expected = 'wsl bootware bootstrap --no-passwd --config ' `
             + '/mnt/c/Users/Administrator/.bootware/config.yaml --inventory ' `
             + '192.48.16.0 --playbook /mnt/c/Fake/path/repo/playbook.yaml ' `
             + "--private-key `$HOME/.ssh/bootware --skip python " `